ref[intro][next-section]: extract components
