--- src/components/organisms/comingSoon/SectionFooter.tsx
@@ -1,10 +1,0 @@
-import Footer from '@/organisms/footer/Footer'
-import SectionTemplate from '@/templates/Section'
-
-export default function SectionFooter() {
-  return (
-    <SectionTemplate mode="light">
-      <Footer />
-    </SectionTemplate>
-  )
-}
