--- src/components/organisms/comingSoon/index.ts
@@ -1,5 +1,0 @@
-export { default as SectionFooter } from './SectionFooter'
-export { default as SectionOne } from './SectionOne'
-export { default as SectionRegister } from './SectionRegister'
-export { default as SectionThree } from './SectionThree'
-export { default as SectionTwo } from './SectionTwo'
