+++ src/components/organisms/executiveSummary/index.ts
@@ -1,3 +1,3 @@
+export { default as ExecutiveSummaryIntro } from '../introSection/IntroSection'
+export { default as NextSection } from '../nextSection/NextSection'
 export { default as Body } from './Body'
-export { default as ExecutiveSummaryIntro } from './IntroSection'
-export { default as NextSection } from './NextSection'
