+++ src/components/organisms/index.ts
@@ -1,6 +1,7 @@
 export { CardArticlePreview } from './card/index'
-export * from './comingSoon/index'
 export { default as Footer } from './footer/Footer'
 export { default as Form } from './form/RegisterForm'
 export { default as Header } from './header/Header'
+export * from './introSection'
 export { ModalRegister } from './modal/index'
+export * from './nextSection'
