+++ src/components/organisms/introSection/IntroDesktop.tsx
@@ -10,16 +10,25 @@ import cn from '@/utils/cn'
 import { MotionDiv } from '@/atoms/index'
 import type { IMotionDiv, IMotionDivPosition, IMotionDivSize } from '@/atoms/motionDiv/motionDiv.type'
 
-import { ASSETS, LANG, MOTION_DIVS_MOBILE, SECTION_2_THRESHOLD } from './introSection.constant'
+import {
+  ASSETS,
+  LANG,
+  MAX_HEIGHT_IN_PX,
+  MAX_WIDTH_IN_PX,
+  MOTION_DIVS_DESKTOP,
+  SECTION_2_THRESHOLD,
+} from './IntroSection.constant'
 import SectionTwo from './SectionTwo'
 
 interface Props {
   handleRegister: () => void
 }
 
-// TODO: reuse the similar code on mobile and desktop
-export default function IntroMobile({ handleRegister }: Readonly<Props>) {
-  const { viewportWidth, viewportHeight, vwToPx } = useViewport({})
+export default function IntroDesktop({ handleRegister }: Readonly<Props>) {
+  const { viewportWidth, viewportHeight, vhToPx, vwToPx } = useViewport({
+    maxHeight: MAX_HEIGHT_IN_PX,
+    maxWidth: MAX_WIDTH_IN_PX,
+  })
   const scrollRef = useRef<HTMLElement>(null)
   const { ref: section2Ref, entry } = useIntersectionObserver({
     threshold: SECTION_2_THRESHOLD,
@@ -27,21 +36,21 @@ export default function IntroMobile({ handleRegister }: Readonly<Props>) {
 
   const motionDivsInPx: { [key: string]: IMotionDiv & { size: IMotionDivSize; position: IMotionDivPosition } } =
     useMemo(() => {
-      return Object.entries(MOTION_DIVS_MOBILE).reduce(
+      return Object.entries(MOTION_DIVS_DESKTOP).reduce(
         (acc: {}, [key, { sizeInPercent, positionInPercent, ...rest }]: [string, IMotionDiv]) => {
           const sizeInPx: IMotionDivSize = {
             w0: vwToPx(sizeInPercent.w0),
-            h0: vwToPx(sizeInPercent.h0),
+            h0: vhToPx(sizeInPercent.h0),
           }
           if (sizeInPercent.w1) sizeInPx.w1 = vwToPx(sizeInPercent.w1)
-          if (sizeInPercent.h1) sizeInPx.h1 = vwToPx(sizeInPercent.h1)
+          if (sizeInPercent.h1) sizeInPx.h1 = vhToPx(sizeInPercent.h1)
 
           const positionInPx: IMotionDivPosition = {
             x0: vwToPx(positionInPercent.x0),
-            y0: vwToPx(positionInPercent.y0),
+            y0: vhToPx(positionInPercent.y0),
           }
           if (positionInPercent.x1) positionInPx.x1 = vwToPx(positionInPercent.x1)
-          if (positionInPercent.y1) positionInPx.y1 = vwToPx(positionInPercent.y1)
+          if (positionInPercent.y1) positionInPx.y1 = vhToPx(positionInPercent.y1)
 
           return { ...acc, [key]: { size: sizeInPx, position: positionInPx, ...rest } }
         },
@@ -55,23 +64,23 @@ export default function IntroMobile({ handleRegister }: Readonly<Props>) {
 
   return (
     <div
-      className={cn('dn-intro--desktop', 'flex flex-col gap-y-[39.77px]', 'w-screen h-full bg-black mt-[75px]')}
+      className={cn('dn-intro--desktop', 'flex flex-col', 'w-screen h-full bg-black', 'overflow-hidden')}
       ref={scrollRef as React.RefObject<HTMLDivElement>}
     >
-      <section className={cn('dn-intro__hero', 'h-min')}>
-        <div className="relative mx-0 h-[100vw] w-full overflow-hidden px-0">
-          {motionDivsInPx.blueMobile && (
+      <section className={cn('dn-intro__hero', 'max-h-[755px] h-screen')}>
+        <div className="container relative mx-auto size-full">
+          {motionDivsInPx.blueDesktop && (
             <MotionDiv
-              {...motionDivsInPx.blueMobile}
+              {...motionDivsInPx.blueDesktop}
               className={cn('bg-accents-blue absolute')}
               motionValue={motionValue}
               inputRange={motionInputRange}
             />
           )}
-          {motionDivsInPx.redPinkMobile && (
+          {motionDivsInPx.redPinkDesktop && (
             <MotionDiv
-              {...motionDivsInPx.redPinkMobile}
-              className={cn('absolute bg-accents-red transition-all')}
+              {...motionDivsInPx.redPinkDesktop}
+              className={cn('absolute bg-accents-red  transition-all')}
               motionValue={motionValue}
               inputRange={motionInputRange}
               ease="easeIn"
@@ -80,24 +89,18 @@ export default function IntroMobile({ handleRegister }: Readonly<Props>) {
             </MotionDiv>
           )}
 
-          <Image
-            fill
-            alt={LANG.heroAlt}
-            src={ASSETS.heroSm}
-            priority
-            className="object-contain transition-opacity delay-300"
-          />
-          {motionDivsInPx.redMobile && (
+          <Image fill alt={LANG.heroAlt} src={ASSETS.heroLg} priority className="transition-opacity delay-300" />
+          {motionDivsInPx.redDesktop && (
             <MotionDiv
-              {...motionDivsInPx.redMobile}
+              {...motionDivsInPx.redDesktop}
               className={cn('absolute bg-accents-red mix-blend-plus-lighter transition-colors')}
               motionValue={motionValue}
               inputRange={motionInputRange}
             />
           )}
-          {motionDivsInPx.cyanGreenMobile && (
+          {motionDivsInPx.cyanGreenDesktop && (
             <MotionDiv
-              {...motionDivsInPx.cyanGreenMobile}
+              {...motionDivsInPx.cyanGreenDesktop}
               className={cn('absolute mix-blend-exclusion transition-all')}
               motionValue={motionValue}
               inputRange={motionInputRange}
@@ -110,7 +113,6 @@ export default function IntroMobile({ handleRegister }: Readonly<Props>) {
           )}
         </div>
       </section>
-      {/* TODO: remove props drilling by bring the section2 component up */}
       <SectionTwo handleRegister={handleRegister} sectionRef={section2Ref} />
     </div>
   )
