+++ src/components/organisms/introSection/IntroSection.constant.ts
