+++ src/components/organisms/introSection/index.ts
@@ -1,0 +1,2 @@
+export { default as IntroSection } from './IntroSection'
+export * from './IntroSection.constant'
