+++ src/components/organisms/nextSection/NextSection.constant.ts
