+++ src/components/organisms/nextSection/index.ts
@@ -1,0 +1,2 @@
+export { default as NextSection } from './NextSection'
+export type { CONTENT as NEXT_SECTION_CONTENT } from './NextSection.constant'
